Move WalletConnect related classes inside its module
