+++ app/src/main/java/io/horizontalsystems/bankwallet/core/storage/AppDatabase.kt
@@ -15,6 +15,8 @@ import io.horizontalsystems.bankwallet.entities.*
 import io.horizontalsystems.bankwallet.modules.nft.NftAssetRecord
 import io.horizontalsystems.bankwallet.modules.nft.NftCollectionRecord
 import io.horizontalsystems.bankwallet.modules.nft.NftDao
+import io.horizontalsystems.bankwallet.modules.walletconnect.entity.WalletConnectSession
+import io.horizontalsystems.bankwallet.modules.walletconnect.storage.WC1SessionDao
 import io.horizontalsystems.marketkit.models.CoinType
 
 @Database(version = 38, exportSchema = false, entities = [
