+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/walletconnect/version1/WC1SessionKillManager.kt
@@ -1,7 +1,7 @@
 package io.horizontalsystems.bankwallet.modules.walletconnect.version1
 
 import io.horizontalsystems.bankwallet.core.managers.WalletConnectInteractor
-import io.horizontalsystems.bankwallet.entities.WalletConnectSession
+import io.horizontalsystems.bankwallet.modules.walletconnect.entity.WalletConnectSession
 import io.reactivex.subjects.PublishSubject
 
 class WC1SessionKillManager(
