+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/walletconnect/version1/WC1SessionManager.kt
@@ -2,10 +2,10 @@ package io.horizontalsystems.bankwallet.modules.walletconnect.version1
 
 import io.horizontalsystems.bankwallet.core.IAccountManager
 import io.horizontalsystems.bankwallet.core.managers.AccountSettingManager
-import io.horizontalsystems.bankwallet.core.storage.WC1SessionStorage
+import io.horizontalsystems.bankwallet.modules.walletconnect.storage.WC1SessionStorage
 import io.horizontalsystems.bankwallet.core.subscribeIO
 import io.horizontalsystems.bankwallet.entities.Account
-import io.horizontalsystems.bankwallet.entities.WalletConnectSession
+import io.horizontalsystems.bankwallet.modules.walletconnect.entity.WalletConnectSession
 import io.reactivex.BackpressureStrategy
 import io.reactivex.Flowable
 import io.reactivex.disposables.CompositeDisposable
